实现了在Invoice中添加实际收费项目并自动计算金额

+++ ClothResorting/Dtos/InvoiceDto.cs
@@ -13,7 +13,7 @@ namespace ClothResorting.Dtos
 
         public string InvoiceType { get; set; }
 
-        public int TotalDue { get; set; }
+        public double TotalDue { get; set; }
 
         public string BillTo { get; set; }
